Retype some Meteor methods
