+++ apps/meteor/definition/methods/index.ts
@@ -1,2 +1,1 @@
 import './federation';
-import './omnichannel';
